# api/api-docs.yaml
{"openapi":"3.0.1","info":{"title":"Smart Document Generation API","description":"This is the backend for the Smart Document Generation Application.","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/document/text":{"post":{"tags":["document-controller"],"operationId":"uploadText","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentTextDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentDTO"}}}}}}},"/api/v1/document/file":{"post":{"tags":["document-controller"],"operationId":"uploadFile","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentDTO"}}}}}}},"/api/v1/assistant/chat":{"post":{"tags":["assistant-controller"],"operationId":"chat","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatMessage"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatMessage"}}}}}}},"/api/v1/document/{id}":{"get":{"tags":["document-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DocumentContentDTO"}}}}}},"delete":{"tags":["document-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/document/download/{id}":{"get":{"tags":["document-controller"],"operationId":"download","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}}},"/api/v1/document/all":{"get":{"tags":["document-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DocumentDTO"}}}}}}}}},"components":{"schemas":{"DocumentTextDTO":{"type":"object","properties":{"content":{"type":"string"},"name":{"type":"string"}}},"DocumentDTO":{"required":["id","name","type"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"type":{"type":"string"},"size":{"type":"integer","format":"int64"},"uploadedAt":{"type":"integer","format":"int64"},"vectorStoreUUIDs":{"type":"array","items":{"type":"string"}}}},"ChatMessage":{"type":"object","properties":{"message":{"type":"string"},"chatId":{"type":"string"},"isResponse":{"type":"boolean"}}},"DocumentContentDTO":{"required":["id","name","type"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"type":{"type":"string"},"size":{"type":"integer","format":"int64"},"uploadedAt":{"type":"integer","format":"int64"},"vectorStoreUUIDs":{"type":"array","items":{"type":"string"}},"content":{"type":"array","items":{"type":"string","format":"byte"}}}}}}}